Move Superglobal to Value/ArrayBase
