+++ tests/value/get_test.php
@@ -1,6 +1,6 @@
 <?php
 
-class SuperglobalGet_Test extends PHPUnit_Framework_TestCase
+class Value_Get_Test extends PHPUnit_Framework_TestCase
 {
     public function testUsesCorrectGlobalVariable()
     {
@@ -9,7 +9,7 @@ class SuperglobalGet_Test extends PHPUnit_Framework_TestCase
             'c' => 'd',
         ];
 
-        $__get = new \Dxw\MyTheme\SuperglobalGet();
+        $__get = new \Dxw\MyTheme\Value\Get();
 
         $this->assertEquals('b', $__get['a']);
         $this->assertEquals('d', $__get['c']);
